+++ CHANGELOG.md
@@ -8,6 +8,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.1.6, 1/15/2024
+
+### Added
+
+Support 3 logging formats using the log.format parameter in application.properties
+
+```
+text - default to text string output
+compact - json output without linefeed for redirection to log analytics system
+json - pretty print for human readers
+```
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 4.1.5, 1/14/2024
 
 ### Added